Adds frontend and broker service
This commit adds the skeleton frontend that will be used to test the
services available. Currently no frontend framework is being used in
favour of vanilla html/css/js to stay flexible.

This commit also impliments the broker service that will be used to
allow the frontend to communicate to the backend as more services are
added.